requirements.txt
tox>=2.5.0
pytest>=3.0.4
twisted
kademlia
gnupg
pycrypto
./pisces
mock
